requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
anthropic==0.42.0
httpx==0.28.1
pyjwt[crypto]==2.10.1
python-dotenv==1.0.1
resend==2.26.0
pydantic==2.10.3
